turret fire

+++ Assets/Scripts/ScriptableObjects/Turrets_Info.cs
@@ -1,12 +1,9 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.UIElements;
 
 [CreateAssetMenu(fileName = "Data", menuName = "ScriptableObjects/Turret")]
 public class Turrets_Info : ScriptableObject
 {
-    public float damage;
+    public int damage;
     public float Range;
     public float fireRate;
     public Sprite model;
